+++ posthog-android/consumer-rules.pro
@@ -23,6 +23,10 @@
 -keep class com.posthog.internal.PostHogDecideResponse { *; }
 -keep class com.posthog.internal.PostHogDecideResponse { <init>(); }
 
+# Session Replay
+-keep class com.posthog.internal.replay.** { *; }
+-keep class com.posthog.internal.replay.** { <init>(); }
+
 # Prevent proguard from stripping interface information from TypeAdapter, TypeAdapterFactory,
 # JsonSerializer, JsonDeserializer instances (so they can be used in @JsonAdapter)
 -keep class * extends com.google.gson.TypeAdapter
