+++ posthog-android/src/main/java/com/posthog/android/replay/internal/ViewTreeSnapshotStatus.kt
@@ -1,10 +1,10 @@
 package com.posthog.android.replay.internal
 
-import com.posthog.internal.RRWireframe
+import com.posthog.internal.replay.RRWireframe
 
 internal class ViewTreeSnapshotStatus(
     val listener: NextDrawListener,
-    var sentFullSnapshot: Boolean = false,
+//    var sentFullSnapshot: Boolean = false,
     var sentMetaEvent: Boolean = false,
     var lastSnapshot: RRWireframe? = null,
 )
