+++ posthog/src/main/java/com/posthog/internal/replay/RRAddedNode.kt
@@ -1,0 +1,9 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRAddedNode(
+    public val wireframe: RRWireframe,
+    public val parentId: Int? = null,
+)
