+++ posthog/src/main/java/com/posthog/internal/replay/RRDocumentNode.kt
@@ -1,0 +1,13 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRDocumentNode(tag: String, payload: Any, timestamp: Long) : RREvent(
+    type = RREventType.Custom,
+    data = mapOf(
+        "tag" to tag,
+        "payload" to payload,
+    ),
+    timestamp = timestamp,
+)
