+++ posthog/src/main/java/com/posthog/internal/replay/RRDomContentLoadedEvent.kt
@@ -1,0 +1,9 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRDomContentLoadedEvent(timestamp: Long) : RREvent(
+    type = RREventType.DomContentLoaded,
+    timestamp = timestamp,
+)
