+++ posthog/src/main/java/com/posthog/internal/replay/RREvent.kt
@@ -1,0 +1,10 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public open class RREvent(
+    public val type: RREventType,
+    public val timestamp: Long,
+    public val data: Any? = null,
+)
