+++ posthog/src/main/java/com/posthog/internal/replay/RRIncrementalMouseInteractionData.kt
@@ -1,0 +1,14 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRIncrementalMouseInteractionData(
+    public val id: Int,
+    public val type: RRMouseInteraction,
+    public val x: Int,
+    public val y: Int,
+    public val source: RRIncrementalSource = RRIncrementalSource.MouseInteraction,
+    public val pointerType: Int = 2, // always Touch
+    public val positions: List<RRMousePosition>? = null,
+)
