+++ posthog/src/main/java/com/posthog/internal/replay/RRLoadedEvent.kt
@@ -1,0 +1,9 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRLoadedEvent(timestamp: Long) : RREvent(
+    type = RREventType.Load,
+    timestamp = timestamp,
+)
