+++ posthog/src/main/java/com/posthog/internal/replay/RRRemovedNode.kt
@@ -1,0 +1,9 @@
+package com.posthog.internal.replay
+
+import com.posthog.PostHogInternal
+
+@PostHogInternal
+public class RRRemovedNode(
+    public val id: Int,
+    public val parentId: Int? = null,
+)
