+++ posthog/src/test/java/com/posthog/internal/PostHogApiTest.kt
@@ -18,8 +18,7 @@ internal class PostHogApiTest {
         host: String,
     ): PostHogApi {
         val config = PostHogConfig(apiKey, host)
-        val dateProvider = PostHogDeviceDateProvider()
-        return PostHogApi(config, dateProvider)
+        return PostHogApi(config)
     }
 
     @Test
