+++ posthog/src/test/java/com/posthog/internal/PostHogFeatureFlagsTest.kt
@@ -33,8 +33,7 @@ internal class PostHogFeatureFlagsTest {
             this.networkStatus = networkStatus
             cachePreferences = preferences
         }
-        val dateProvider = PostHogDeviceDateProvider()
-        val api = PostHogApi(config, dateProvider)
+        val api = PostHogApi(config)
         return PostHogFeatureFlags(config, api, executor = executor)
     }
